% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_curtailed.R
\name{build_curtailed}
\alias{build_curtailed}
\title{Build a curtailed group sequential single-arm trial design}
\usage{
build_curtailed(
  J = 2,
  n = c(10, 19),
  a = c(1, 5),
  r = c(Inf, 6),
  a_curt,
  r_curt,
  pi0 = 0.1,
  pi1 = 0.3,
  alpha = 0.05,
  beta = 0.2,
  summary = TRUE
)
}
\arguments{
\item{J}{The maximal number of stages. Must be a positive integer.}

\item{n}{A vector of stage sample sizes of length \code{J}.}

\item{a}{A vector of futility boundaries of length \code{J}. Use
\code{-Inf} to suppress early stopping for futility at a given stage.}

\item{r}{A vector of efficacy boundaries of length \code{J}. Use
\code{Inf} to suppress early stopping for efficacy at a given stage.}

\item{a_curt}{A vector of curtailed futility boundaries of length
\code{sum(n)}.}

\item{r_curt}{A vector of curtailed efficacy boundaries of length
\code{sum(n)}.}

\item{pi0}{The (undesirable) response probability used in the definition of
the null hypothesis.}

\item{pi1}{The (desirable) response probability at which the trial is
powered.}

\item{alpha}{The desired maximal type-I error-rate.}

\item{beta}{The desired maximal type-II error-rate.}

\item{summary}{A logical variable indicating whether a summary of the
function's progress should be printed to the console.}
}
\value{
A list of class \code{"sa_des_curtailed"} containing the following
  elements
  \itemize{
    \item A list in the slot \code{$des} containing the design parameters.
    \item Each of the input variables as specified.
  }
}
\description{
Constructs an object of class \code{"sa_des_curtailed"} from user-supplied
design parameters, without performing an optimisation search. This allows
bespoke curtailed group sequential designs to be passed to downstream
functions such as \code{\link{opchar_curtailed}} and
\code{\link{est_curtailed}}.
}
\examples{
des <- build_curtailed(J = 2, n = c(10, 19), a = c(1, 5), r = c(Inf, 6),
                       a_curt = c(rep(-Inf, 28), 5),
                       r_curt = c(rep(Inf, 28), 6))
}
\seealso{
\code{\link{des_curtailed}}, \code{\link{opchar_curtailed}},
  \code{\link{est_curtailed}}, and their associated \code{plot} family of
  functions.
}
