% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gs.R
\name{build_gs}
\alias{build_gs}
\title{Build a group sequential single-arm trial design}
\usage{
build_gs(
  J = 2,
  n = c(10, 19),
  a = c(1, 5),
  r = c(Inf, 6),
  pi0 = 0.1,
  pi1 = 0.3,
  alpha = 0.05,
  beta = 0.2,
  summary = TRUE
)
}
\arguments{
\item{J}{The maximal number of stages. Must be a positive integer.}

\item{n}{A vector of stage sample sizes of length \code{J}.}

\item{a}{A vector of futility boundaries of length \code{J}. Use
\code{-Inf} to suppress early stopping for futility at a given stage.}

\item{r}{A vector of efficacy boundaries of length \code{J}. Use
\code{Inf} to suppress early stopping for efficacy at a given stage.}

\item{pi0}{The (undesirable) response probability used in the definition of
the null hypothesis.}

\item{pi1}{The (desirable) response probability at which the trial is
powered.}

\item{alpha}{The desired maximal type-I error-rate.}

\item{beta}{The desired maximal type-II error-rate.}

\item{summary}{A logical variable indicating whether a summary of the
function's progress should be printed to the console.}
}
\value{
A list of class \code{"sa_des_gs"} containing the following
  elements
  \itemize{
    \item A list in the slot \code{$des} containing the design parameters.
    \item Each of the input variables as specified.
  }
}
\description{
Constructs an object of class \code{"sa_des_gs"} from user-supplied design
parameters, without performing an optimisation search. This allows bespoke
group sequential designs to be passed to downstream functions such as
\code{\link{opchar_gs}}, \code{\link{est_gs}}, \code{\link{pval_gs}}, and
\code{\link{ci_gs}}.
}
\examples{
des <- build_gs(J = 2, n = c(10, 19), a = c(1, 5), r = c(Inf, 6))
}
\seealso{
\code{\link{des_gs}}, \code{\link{opchar_gs}},
  \code{\link{est_gs}}, \code{\link{pval_gs}}, \code{\link{ci_gs}},
  and their associated \code{plot} family of functions.
}
