% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_fixed.R
\name{ci_fixed}
\alias{ci_fixed}
\title{Determine confidence intervals in a single-stage single-arm trial design for
a single binary endpoint}
\usage{
ci_fixed(des, pi, alpha = des$alpha, method = c("agresti_coull",
  "clopper_pearson", "jeffreys", "mid_p", "wald", "wilson"), summary = F)
}
\arguments{
\item{des}{An object of class \code{"sa_des_fixed"}, as returned by
\code{des_fixed()}.}

\item{pi}{A vector of response probabilities to evaluate the expected
performance of the confidence interval calculation procedure at. This will
internally default to be the
\ifelse{html}{\out{<i>&pi;</i><sub>0</sub>}}{\deqn{\pi_0}} and
\ifelse{html}{\out{<i>&pi;</i><sub>1</sub>}}{\deqn{\pi_1}} from \code{des} if
it is left unspecified.}

\item{alpha}{\ifelse{html}{\out{<i>&alpha;</i>}}{\deqn{\alpha}}-level to use
in confidence interval construction. Defaults to the value of
\ifelse{html}{\out{<i>&alpha;</i>}}{\deqn{\alpha}} used in the construction
of \code{des} (i.e., \code{des$alpha}).}

\item{method}{A vector of methods to use to construct p-values. Currently,
support is available to use the Agresti-Coull (\code{"agresti_coull"}),
Clopper-Pearson (\code{"clopper_pearson"}), Jeffreys (\code{"jeffreys"}),
Mid-p (\code{"mid_p"}), Wald (\code{"wald"}), and Wilson Score
(\code{"wilson"}) approaches to confidence interval determination.}

\item{summary}{A logical variable indicating whether a summary of the
function's progress should be printed to the console.}
}
\value{
A list of class \code{"sa_ci_fixed"} containing the following
elements
\itemize{
\item A tibble in the slot \code{$ci} summarising the possible confidence
intervals at the end of the trial for the supplied design, according to the
chosen methods.
\item A tibble in the slot \code{$perf} summarising the performance of the
chosen confidence interval determination procedures for each specified value
of \ifelse{html}{\out{<i>&pi;</i>}}{\deqn{\pi}}.
\item Each of the input variables as specified, subject to internal
modification.
}
}
\description{
Determines all possible confidence intervals at the end of a single-stage
single-arm trial for a single binary endpoint, as determined
using \code{des_fixed()}. Support is available to compute confidence
intervals using the Agresti-Coull, Clopper-Pearson, Jeffreys, Mid-p, Wald,
and Wilson Score approaches.
}
\details{
In addition, the performance of the chosen confidence interval determination
procedures (including their coverage and expected length) for each value of
\ifelse{html}{\out{<i>pi</i>}}{\eqn{\pi}} in the supplied vector
\ifelse{html}{\out{<b><i>pi</i></b>}}{\eqn{\bold{\pi}}}, will also be
evaluated.
}
\examples{
# Find the optimal single-stage design for the default parameters
des <- des_fixed()
# Determine the performance of all supported confidence interval
# determination procedures for a range of possible response probabilities
ci  <- ci_fixed(des, pi = seq(0, 1, 0.01))
}
\seealso{
\code{\link{des_fixed}}, \code{\link{opchar_fixed}},
\code{\link{est_fixed}}, \code{\link{pval_fixed}}, and their associated
\code{plot} family of functions.
}
