% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_gehan.R
\name{ci_gehan}
\alias{ci_gehan}
\title{Determine confidence intervals in gehan single-arm trial designs
for a single binary endpoint}
\usage{
ci_gehan(des, k, pi, alpha = des$alpha, summary = F)
}
\arguments{
\item{des}{An object of class \code{"sa_des_gehan"}, as returned by
\code{des_gehan()}.}

\item{k}{Calculations are performed conditional on the trial stopping in one
of the stages listed in vector \code{k}. Thus, \code{k} should be a vector of
integers, with elements between one and the maximum number of possible
stages.}

\item{pi}{A vector of response probabilities to evaluate the expected
performance of the point estimation procedure at. This will internally
default to be the \ifelse{html}{\out{<i>&pi;</i><sub>0</sub>}}{\deqn{\pi_0}}
and \ifelse{html}{\out{<i>&pi;</i><sub>1</sub>}}{\deqn{\pi_1}} from
\code{des} if it is left unspecified.}

\item{alpha}{Level to use in confidence interval construction. Defaults to
the value of \ifelse{html}{\out{<i>&alpha;</i>}}{\deqn{\alpha}} used in the
construction of \code{des} (i.e., \code{des$alpha}).}

\item{summary}{A logical variable indicating whether a summary of the
function's progress should be printed to the console.}
}
\value{
A list of class \code{"sa_ci_gehan"} containing the following
elements
\itemize{
\item A tibble in the slot \code{$ci} summarising the possible confidence
intervals at the end of the trial for the supplied design, according to the
chosen methods.
\item A tibble in the slot \code{$perf} summarising the performance of the
chosen confidence interval determination procedures for each specified value
of \ifelse{html}{\out{<i>&pi;</i>}}{\deqn{\pi}}.
\item Each of the input variables as specified, subject to internal
modification.
}
}
\description{
Determines possible confidence intervals at the end of a gehan
single-arm trial for a single binary endpoint, as determined
using \code{des_gehan()}. Support is available to compute confidence
intervals a naive (\code{"naive"}) approach only.
}
\details{
In addition, the performance of the chosen confidence interval determination
procedures (including their coverage and expected length) for each value of
\ifelse{html}{\out{<i>pi</i>)}}{\eqn{\pi}} in the supplied vector
\ifelse{html}{\out{<b><i>pi</i></b>)}}{\eqn{\bold{\pi}}}, will also be
evaluated.

Calculations are performed conditional on the trial stopping in one of the
stages specified using the input (vector) \code{k}.
}
\examples{
# Find the gehan two-stage design for the default parameters
des <- des_gehan()
# Determine the performance of all supported confidence interval
# determination procedures for a range of possible response probabilities
ci  <- ci_gehan(des, pi = seq(0, 1, 0.01))
}
\seealso{
\code{\link{des_gehan}} and \code{\link{opchar_gehan}}.
}
