% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_gs.R
\name{ci_gs}
\alias{ci_gs}
\title{Determine confidence intervals in a group sequential single-arm trial design
for a single binary endpoint}
\usage{
ci_gs(des, k, pi, alpha = des$alpha, method = c("exact", "mid_p",
  "naive"), summary = F)
}
\arguments{
\item{des}{An object of class \code{"sa_des_gs"}, as returned by
\code{des_gs()}.}

\item{k}{Calculations are performed conditional on the trial stopping in one
of the stages listed in vector \code{k}. Thus, \code{k} should be a vector of
integers, with elements between one and the maximum number of possible
stages.}

\item{pi}{A vector of response probabilities to evaluate the expected
performance of the point estimation procedures at. This will internally
default to be the \ifelse{html}{\out{<i>&pi;</i><sub>0</sub>}}{\deqn{\pi_0}}
and \ifelse{html}{\out{<i>&pi;</i><sub>1</sub>}}{\deqn{\pi_1}} from
\code{des} if it is left unspecified.}

\item{alpha}{Level to use in confidence interval construction. Defaults to
the value of \ifelse{html}{\out{<i>&alpha;</i>}}{\deqn{\alpha}} used in the
construction of \code{des} (i.e., \code{des$alpha}).}

\item{method}{A vector of methods to use to construct confidence intervals.
Currently, support is available to use the naive (\code{"naive"}), exact
(\code{"exact"}), and Mid-p (\code{"mid_p"}) approaches. Defaults to all
available methods.}

\item{summary}{A logical variable indicating whether a summary of the
function's progress should be printed to the console.}
}
\value{
A list of class \code{"sa_ci_gs"} containing the following elements
\itemize{
\item A tibble in the slot \code{$ci} summarising the possible confidence
intervals at the end of the trial for the supplied design, according to the
chosen methods.
\item A tibble in the slot \code{$perf} summarising the performance of the
chosen confidence interval determination procedures for each specified value
of \ifelse{html}{\out{<i>&pi;</i>}}{\deqn{\pi}}.
\item Each of the input variables as specified, subject to internal
modification.
}
}
\description{
Determines possible confidence intervals at the end of a
group sequential single-arm trial for a single binary endpoint, as determined
using \code{des_gs()}. Support is available to compute confidence intervals
using the naive (\code{"naive"}), exact (\code{"exact"}), and Mid-p
(\code{"mid_p"}) approaches.
}
\details{
In addition, the performance of the chosen confidence interval determination
procedures (including their coverage and expected length) for each value of
\ifelse{html}{\out{<i>pi</i>)}}{\eqn{\pi}} in the supplied vector
\ifelse{html}{\out{<b><i>pi</i></b>)}}{\eqn{\bold{\pi}}}, will also be
evaluated.

Calculations are performed conditional on the trial stopping in one of the
stages specified using the input (vector) \code{k}.
}
\examples{
# Find the optimal two-stage design for the default parameters
des <- des_gs()
# Determine the performance of all supported confidence interval
# determination procedures for a range of possible response probabilities
ci  <- ci_gs(des, pi = seq(0, 1, 0.01))
}
\seealso{
\code{\link{des_gs}}, \code{\link{opchar_gs}}, \code{\link{est_gs}},
\code{\link{pval_gs}}, and their associated \code{plot} family of functions.
}
