% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_bayesfreq.R
\name{des_bayesfreq}
\alias{des_bayesfreq}
\title{Design a Bayesian-frequentist single-arm trial for a single binary endpoint}
\usage{
des_bayesfreq(J = 2, pi0 = 0.1, pi1 = 0.3, alpha = 0.05, beta = 0.2,
  mu = 0.1, nu = 0.9, Nmin = 1, Nmax = 30, optimality = "ess",
  control = c("frequentist", "bayesian"), equal_n = F, PL = 0.5,
  PU = 0.9, PT = 0.95, summary = F)
}
\arguments{
\item{J}{The maximal number of stages to allow.}

\item{pi0}{The (undesirable) response probability used in the definition of
the null hypothesis.}

\item{pi1}{The (desiable) response probability used in the definition of the
alternative hypothesis.}

\item{alpha}{The desired maximal type-I error-rate.}

\item{beta}{The desired maximal type-II error-rate.}

\item{mu}{The first shape parameter of the Beta distribution.}

\item{nu}{The second shape parameter of the Beta distribution.}

\item{Nmin}{The minimal total sample size to allow in considered
designs.}

\item{Nmax}{The maximal total sample size to allow in considered designs.}

\item{optimality}{Choice of optimal design criteria. Must be one of
\code{"ess"} or \code{"minimax"}.}

\item{control}{Error-rates to control. Should be a vector containing elements
chosen from "frequentist" and "bayesian".}

\item{equal_n}{A logical variable indicating that the sample size of each
stage should be equal.}

\item{PL}{Predictive probability value used in determining when to stop the
trial early for futility.}

\item{PU}{Predictive probability value used in determining when to stop the
trial early for futility.}

\item{PT}{Terminal predictie probability value used in determining when the
trial is a success.}

\item{summary}{A logical variable indicating a summary of the function's
progress should be printed to the console.}
}
\value{
A list of class \code{"sa_des_bayesfreq"} containing the following
elements
\itemize{
\item A list in the slot \code{$des} containing details of the
identified optimal design.
\item A tibble in the slot \code{$feasible}, consisting of the
identified designs which met the required operating characteristics.
\item Each of the input variables as specified.
}
}
\description{
Determines optimised single- and two-stage Bayesian-frequentst single-arm
clinical trial designs for a single binary primary endpoint, using exact
calculations.
}
\details{
Designs controlling Bayesian, frequentist, or Bayesian and frequentist
operating characteristics can be determining, which optimise either the
Bayesian expected sample size or the maximal sample size.
}
\examples{
# The ESS-optimal design for the default parameters
ess_optimal <- des_bayesfreq()
# The corresponding minimax design
minimax     <- des_bayesfreq(optimality = "minimax")
}
\seealso{
\code{\link{opchar_bayesfreq}}, and their associated \code{plot}
family of functions.
}
