% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_fixed.R
\name{des_fixed}
\alias{des_fixed}
\title{Design a single-stage single-arm trial for a single binary endpoint}
\usage{
des_fixed(pi0 = 0.1, pi1 = 0.3, alpha = 0.05, beta = 0.2, Nmin = 1,
  Nmax = 30, exact = T, summary = F)
}
\arguments{
\item{pi0}{The (undesirable) response probability used in the definition of
the null hypothesis. Defaults to 0.1.}

\item{pi1}{The (desirable) response probability at which the trial is
powered. Defaults to 0.3.}

\item{alpha}{The desired maximal type-I error rate. Defaults to 0.05.}

\item{beta}{The desired maximal type-II error rate. Defaults to 0.2.}

\item{Nmin}{The minimal sample size to allow in considered designs. Defaults
to 1.}

\item{Nmax}{The maximal sample size to allow in considered designs. Defaults
to 30.}

\item{exact}{A logical variable indicating whether exact binomial
calculations or a normal approximation approach should be used to determine
the optimal design. Defaults to TRUE.}

\item{summary}{A logical variable indicating whether a summary of the
function's progress should be printed to the console. Defaults to FALSE.}
}
\value{
A list of class \code{"sa_des_fixed"} containing the following
elements
\itemize{
\item A tibble in the slot \code{$des} summarising the characteristics of the
identified optimal design. This will be \code{NULL} if no feasible design was
identified in the considered range for
\ifelse{html}{\out{<i>n</i>}}{\eqn{n}}.
\item A tibble in the slot \code{$feasible}, consisting of the identified
designs which met the required operating characteristics.
\item Each of the input variables as specified, subject to internal
modification.
}
}
\description{
Determines single-stage single-arm clinical trial designs
for a single binary primary endpoint, using either exact binomial
calculations, or a normal approximation approach.
}
\details{
\code{des_fixed()} supports the determination of single-stage single-arm
clinical trial designs for a single binary primary endpoint. The following
hypotheses are tested for the response probability
\ifelse{html}{\out{<i>&pi;</i>}}{\eqn{\pi}}

\ifelse{html}{\out{<center><i>H</i><sub>0</sub> : <i>&pi;</i> &le; <i>&pi;
</i><sub>0</sub>, <i>H</i><sub>1</sub> : <i>&pi;</i> > <i>&pi;</i><sub>
0</sub>,</center>}}{\deqn{H_0 : \pi \le \pi_0,\qquad H_1 : \pi > \pi_0,}}

for \ifelse{html}{\out{<i>&pi;</i><sub>0</sub>}}{\eqn{\pi_0}},  specified
using the argument \code{pi0}.

In each instance, the optimal design is required to meet the following
operating characteristics

\ifelse{html}{\out{<center><i>P</i>(<i>&pi;</i><sub>0</sub>) &le;
<i>&alpha;</i>, <i>P</i>(<i>&pi;</i><sub>1</sub>) &ge; 1 - <i>&beta;</i>,
</center>}}{\deqn{P(\pi_0) \le \alpha,\qquad P(\pi_1) \ge 1 - \beta,}}

where \ifelse{html}{\out{<i>P</i>(<i>&pi;</i>)}}{\eqn{P(\pi)}} is the
probability of rejecting \ifelse{html}{\out{<i>H</i><sub>0</sub>}}{\eqn{H_0}}
when the true response probability is
\ifelse{html}{\out{<i>&pi;</i>}}{\eqn{\pi}}, and the values of
\ifelse{html}{\out{<i>&alpha;</i>}}{\eqn{\alpha}} and
\ifelse{html}{\out{<i>&beta;</i>}}{\eqn{\beta}} are specified using the
arguments \code{alpha} and \code{beta} respectively. Moreover,
\ifelse{html}{\out{<i>&pi;</i><sub>1</sub>}}{\eqn{\pi_1}}, satisfying
\ifelse{html}{\out{<i>&pi;</i><sub>0</sub> &lt;
<i>&pi;</i><sub>1</sub>}}{\eqn{\pi_0 < \pi_1}}, is specified using the
argument \code{pi1}.

A single-stage single-arm design for a single binary endpoint is ultimately
indexed by three parameters: \ifelse{html}{\out{<i>a</i>}}{\eqn{a}},
\ifelse{html}{\out{<i>r</i>}}{\eqn{r}}, and
\ifelse{html}{\out{<i>n</i>}}{\eqn{n}}.

With these parameters, and denoting the number of responses after
\ifelse{html}{\out{<i>m</i>}}{\eqn{m}} outcomes have been observed by
\ifelse{html}{\out{<i>s</i><sub><i>m</i></sub>}}{\eqn{s_m}}, the testing
rules for the trial are as follows

\itemize{
\item If \ifelse{html}{\out{<i>s</i><sub><i>n</i></sub>
&le; <i>a</i>}}{\eqn{s_{n} \le a}}, then do not reject
\ifelse{html}{\out{<i>H</i><sub>0</sub>}}{\eqn{H_0}}.
\item Else if \ifelse{html}{\out{<i>s</i><sub><i>n</i></sub>
&ge; <i>r</i>}}{\eqn{s_{n} \ge r}}, then reject
\ifelse{html}{\out{<i>H</i><sub>0</sub>}}{\eqn{H_0}}.
}

The purpose of this function is then to determine (optimise)
\ifelse{html}{\out{<i>a</i>}}{\eqn{a}},
\ifelse{html}{\out{<i>r</i>}}{\eqn{r}}, and
\ifelse{html}{\out{<i>n</i>}}{\eqn{n}}, accounting for the chosen
restrictions placed on these parameters.

The arguments \code{Nmin} and \code{Nmax} allow restrictions
to be placed on \ifelse{html}{\out{<i>n</i>}}{\eqn{n}}.
Precisely, \code{Nmin} and \code{Nmax} set an inclusive range of allowed
values for \ifelse{html}{\out{<i>n</i>}}{\eqn{n}}.

Note that to ensure a decision is made about
\ifelse{html}{\out{<i>H</i><sub>0</sub>}}{\eqn{H_0}}, this function always
enforces \ifelse{html}{\out{<i>a</i> + 1 = <i>r</i>}}{\eqn{a + 1 = r}}.

The optimal design is then the one that minimises
\ifelse{html}{\out{<i>n</i>}}{\eqn{n}}. In the case where there are multiple
feasible designs with the same minimal value of
\ifelse{html}{\out{<i>n</i>}}{\eqn{n}}, the optimal design is the one amongst
these which maximises
\ifelse{html}{\out{<i>P</i>(<i>&pi;</i><sub>1</sub>)}}{\eqn{P(\pi_1)}}.

If \code{exact} is set to \code{TRUE} then exact binomial probability
calculations are used to identify the optimal design. Otherwise, a normal
approximation approach is used. Note that setting \code{exact = TRUE} is
recommended.
}
\examples{
# The optimal design for the default parameters
des    <- des_fixed()
# The optimal single-stage design for a 10\% type-I error rate
des_10 <- des_fixed(alpha = 0.1)
}
\references{
A'Hern RP (2001) Sample size tables for exact single-stage phase
II designs. \emph{Statistics in Medicine} \strong{20:}859-66.

Fleming TR (1982) One-sample multiple testing procedure for phase
II clinical trials. \emph{Biometrics} \strong{38:}143-51.
}
\seealso{
\code{\link{opchar_fixed}}, \code{\link{est_fixed}},
\code{\link{pval_fixed}}, \code{\link{ci_fixed}}, and their
associated \code{plot} family of functions. Note that similar functionality
is available through \code{\link[clinfun]{ph2single}}.
}
