% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_gehan.R
\name{des_gehan}
\alias{des_gehan}
\title{Design a two-stage Gehan single-arm trial for a single binary endpoint}
\usage{
des_gehan(pi1 = 0.3, beta1 = 0.1, gamma = 0.05,
  alpha_pi_hat = 0.25, method = "A", f = "G", alpha = 0.05,
  pi0 = 0.1, find_D = F, summary = F)
}
\arguments{
\item{pi1}{The (desirable) response probability used in computing the first
stage sample size.}

\item{beta1}{The desired maximal type-II error-rate for stage one.}

\item{gamma}{The desired standard error for the estimate of the response
probability by the end of stage two.}

\item{alpha_pi_hat}{The confidence level to use in the formula for computing
the interim estimated of the response rate.}

\item{method}{Should be either "A" or "B", signifying the method used for
computing the interim response rate.}

\item{f}{Should be either "G", "L", or "EL", signifying the function used for
determining the second stage sample sizes.}

\item{alpha}{The confidence level to use in the formula for computing the
second stage sample sizes.}

\item{pi0}{The null response rate. Only used when \code{find_D = T}.}

\item{find_D}{A logical variable indicating whether an optimal discrete
conditional error function should be found.}

\item{summary}{A logical variable indicating a summary of the function's
progress should be printed to the console.}
}
\value{
A list of class \code{"sa_des_gehan"} containing the following
elements
\itemize{
\item A list in the slot \code{$des} containing details of the identified
optimal design.
\item A tibble in the slot \code{$feasible}, consisting of the
identified designs which met the required operating characteristics.
\item Each of the input variables as specified.
}
}
\description{
Determines two-stage Gehan single-arm clinical trial designs for a single
binary primary endpoint.
}
\details{
\code{des_gehan()} supports the determination of two-stage Gehan
single-arm clinical trial designs for a single binary primary endpoint.
}
\examples{
# The default design
gehan <- des_gehan()
}
