% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_curtailed.R
\name{est_curtailed}
\alias{est_curtailed}
\title{Determine point estimates in a curtailed group sequential single-arm trial
design for a single binary endpoint}
\usage{
est_curtailed(des, k, pi, method = c("bias_adj", "bias_sub",
  "conditional", "naive", "mue", "umvue"), summary = F)
}
\arguments{
\item{des}{An object of class \code{"sa_des_curtailed"}, as returned by
\code{des_curtailed()}.}

\item{k}{Calculations are performed conditional on the trial stopping in one
of the stages listed in vector \code{k}. Thus, \code{k} should be a vector of
integers, with elements between one and the maximum number of possible
stages. If left unspecified, it will internally default to all possible
stages.}

\item{pi}{A vector of response probabilities to evaluate the expected
performance of the point estimation procedures at. This will internally
default to be the \ifelse{html}{\out{<i>&pi;</i><sub>0</sub>}}{\deqn{\pi_0}}
and \ifelse{html}{\out{<i>&pi;</i><sub>1</sub>}}{\deqn{\pi_1}} from
\code{des} if it is left unspecified.}

\item{method}{A vector of methods to use to construct point estimates.
Currently, support is available to use the naive (\code{"naive"}),
bias-adjusted (\code{"bias_adj"}), bias-subtracted (\code{"bias_sub"}),
conditional (\code{"conditional"}), median unbiased (\code{"mue"}), and UMVUE
(\code{"umvue"}) approaches. Defaults to all available methods.}

\item{summary}{A logical variable indicating whether a summary of the
function's progress should be printed to the console.}
}
\value{
A list of class \code{"sa_est_curtailed"} containing the following
elements
\itemize{
\item A tibble in the slot \code{$est} summarising the possible point
estimates at the end of the trial for the supplied design, according to the
chosen methods.
\item A tibble in the slot \code{$perf} summarising the performance of the
chosen point estimation procedures for each specified value of
\ifelse{html}{\out{<i>&pi;</i>}}{\deqn{\pi}}.
\item Each of the input variables as specified, subject to internal
modification.
}
}
\description{
Determines possible point estimates at the end of a curtailed group
sequential single-arm trial for a single binary endpoint, as determined using
\code{des_curtailed()}. Support is available to compute point estimates using
the naive (\code{"naive"}), bias-adjusted (\code{"bias_adj"}),
bias-subtracted (\code{"bias_sub"}), conditional (\code{"conditional"}),
median unbiased (\code{"mue"}), and UMVUE (\code{"umvue"}) approaches.
}
\details{
In addition, the performance of the chosen point estimate procedures
(including their expected value and variance) for each value of
\ifelse{html}{\out{<i>pi</i>}}{\eqn{\pi}} in the supplied vector
\ifelse{html}{\out{<b><i>pi</i></b>}}{\eqn{\bold{\pi}}}, will also be
evaluated.

Calculations are performed conditional on the trial stopping in one of the
stages specified using the input (vector) \code{k}.
}
\examples{
# Find the optimal non-stochastically curtailed two-stage design for the
default parameters
des <- des_curtailed()
# Determine the performance of all supported point estimation procedures for
# a range of possible response probabilities
est <- est_gs(des, pi = seq(0, 1, 0.01))
}
\seealso{
\code{\link{des_curtailed}}, \code{\link{opchar_curtailed}},
\code{\link{pval_curtailed}}, \code{\link{ci_curtailed}}, and their
associated \code{plot} family of functions.
}
