% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opchar_admissable.R
\name{opchar_admissable}
\alias{opchar_admissable}
\title{Determine the operating characteristics of admissable group sequential
single-arm trial designs for a single binary endpoint}
\usage{
opchar_admissable(des, k, pi, summary = F)
}
\arguments{
\item{des}{An object of class \code{"sa_des_admissable"}, as returned by
\code{des_admissable()}.}

\item{k}{Calculations are performed conditional on the trial stopping in one
of the stages listed in vector \code{k}. Thus, \code{k} should be a vector of
integers, with elements between one and the maximal number of possible stages
in the supplied designs. If left unspecified, it will internally default to
all possible stages.}

\item{pi}{A vector of response probabilities to evaluate operating
characteristics at. This will internally default to be the
\ifelse{html}{\out{<i>&pi;</i><sub>0</sub>}}{\deqn{\pi_0}} and
\ifelse{html}{\out{<i>&pi;</i><sub>1</sub>}}{\deqn{\pi_1}} from the
supplied designs if it is left unspecified.}

\item{summary}{A logical variable indicating whether a summary of the
unction's progress should be printed to the console.}
}
\value{
A list of class \code{"sa_opchar_admissable"} containing the
following elements
\itemize{
\item A tibble in the slot \code{$opchar} summarising the operating
characteristics of the supplied designs.
\item Each of the input variables as specified, subject to internal
modification.
}
}
\description{
\code{opchar_admissable()} supports the evaluation of the operating
characteristics of admissable group sequential single-arm clinical trial
designs for a single binary primary endpoint, determined using
\code{des_admissable()}.
}
\details{
For each value of \ifelse{html}{\out{<i>pi</i>}}{\eqn{\pi}} in
the supplied vector \ifelse{html}{\out{<b><i>pi</i></b>}}{\eqn{\bold{\pi}}},
\code{opchar_admissable()} evaluates the power, ESS, and other key
characteristics, of each of the supplied designs.

Calculations are performed conditional on the trial stopping in one of the
stages specified using the input (vector) \code{k}.
}
\examples{
# Find the admissable two-stage design for the default parameters
des         <- des_admissable()
# Determine operating characteristics for a range of response probabilities
opchar      <- opchar_admissable(des, pi = seq(0, 1, 0.01))
}
\seealso{
\code{\link{des_admissable}}, and their associated \code{plot}
family of functions.
}
