% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opchar_fixed.R
\name{opchar_fixed}
\alias{opchar_fixed}
\title{Determine the operating characteristics of single-stage single-arm trial
designs for a single binary endpoint}
\usage{
opchar_fixed(des, ..., pi, summary = F)
}
\arguments{
\item{des}{An object of class \code{"sa_des_fixed"}, as returned by
\code{des_fixed()}.}

\item{...}{Additional objects of class \code{"sa_des_fixed"}. These will be
grouped in to a list named \code{"add_des"}.}

\item{pi}{A vector of response probabilities to evaluate operating
characteristics at. This will internally default to be
\ifelse{html}{\out{<i>&pi;</i><sub>0</sub>}}{\deqn{\pi_0}} and the
\ifelse{html}{\out{<i>&pi;</i><sub>1</sub>}}{\deqn{\pi_1}} from the supplied
designs if it is left unspecified.}

\item{summary}{A logical variable indicating whether a summary of the
function's progress should be printed to the console.}
}
\value{
A list of class \code{"sa_opchar_fixed"} containing the following
elements
\itemize{
\item A tibble in the slot \code{$opchar} summarising the operating
characteristics of the supplied designs. The power for each
\ifelse{html}{\out{<i>&pi;</i>}}{\deqn{\pi}} will be of principle use, but
other characteristics are also provided for convenience.
\item A tibble in the slot \code{$pmf} containing details of the PMF of each
of the supplied designs.
\item Each of the input variables as specified, subject to internal
modification.
}
}
\description{
\code{opchar_fixed()} supports the simultaneous evaluation of the operating
characteristics of multiple single-stage single-arm clinical trial designs
for a single binary primary endpoint, determined using \code{des_fixed()}.
}
\details{
Note that each of the supplied designs must have been designed for the same
value of \ifelse{html}{\out{<i>&pi;</i><sub>0</sub>}}{\deqn{\pi_0}}.

For each value of \ifelse{html}{\out{<i>&pi;</i>}}{\eqn{\pi}} in
the supplied vector
\ifelse{html}{\out{<b><i>&pi;</i></b>}}{\eqn{\bold{\pi}}} (\code{pi}),
\code{opchar_fixed()} evaluates the power of each of the supplied designs.
Additional operating characteristics are also returned for convenience,
specifically to simplify combining the output with that from other
\code{opchar_###()} functions.
}
\examples{
# Find the optimal single-stage design for the default parameters
des         <- des_fixed()
# Determine operating characteristics for a range of response probabilities
opchar      <- opchar_fixed(des, pi = seq(0, 1, 0.01))
# Find the optimal single-stage design for a 10\% type-I error rate
des_10      <- des_fixed(alpha = 0.1)
# Determine operating characteristics for both designs at the same time
opchar_both <- opchar_fixed(des, des_10,
                            pi = seq(0, 1, 0.01))
}
\references{
A'Hern RP (2001) Sample size tables for exact single-stage phase
II designs.  \emph{Statistics in Medicine} \strong{20:}859-66.

Fleming TR (1982) One-sample multiple testing procedure for phase
II clinical trials. \emph{Biometrics} \strong{38:}143-51.
}
\seealso{
\code{\link{des_fixed}}, \code{\link{est_fixed}},
\code{\link{pval_fixed}}, \code{\link{ci_fixed}}, and their associated
\code{plot} family of functions.
}
