% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opchar_gs.R
\name{opchar_gs}
\alias{opchar_gs}
\title{Determine the operating characteristics of group sequential single-arm trial
designs for a single binary endpoint}
\usage{
opchar_gs(des, ..., k, pi, summary = F)
}
\arguments{
\item{des}{An object of class \code{"sa_des_gs"}, as returned by
\code{des_gs()}.}

\item{...}{Additional objects of class \code{"sa_gs_gs"}. These will be
grouped in to a list named \code{"add_des"}.}

\item{k}{Calculations are performed conditional on the trial stopping in one
of the stages listed in vector \code{k}. Thus, \code{k} should be a vector of
integers, with elements between one and the maximum number of possible stages
across the supplied designs. If left unspecified, it will internally default
to all possible stages.}

\item{pi}{A vector of response probabilities to evaluate operating
characteristics at. This will internally default to be the
\ifelse{html}{\out{<i>&pi;</i><sub>0</sub>}}{\deqn{\pi_0}} and
\ifelse{html}{\out{<i>&pi;</i><sub>1</sub>}}{\deqn{\pi_1}} from the
supplied designs if it is left unspecified.}

\item{summary}{A logical variable indicating whether a summary of the
function's progress should be printed to the console.}
}
\value{
A list of class \code{"sa_opchar_gs"} containing the following
elements
\itemize{
\item A tibble in the slot \code{$opchar} summarising the operating
characteristics of the supplied designs.
\item Each of the input variables as specified, subject to internal
modification.
}
}
\description{
\code{opchar_gs()} supports the simultaneous evaluation of the operating
characteristics of multiple group sequential single-arm clinical trial
designs for a single binary primary endpoint, determined using
\code{des_gs()}.
}
\details{
Note that each of the supplied designs must have been designed for the same
value of \ifelse{html}{\out{<i>&pi;</i><sub>0</sub>}}{\deqn{\pi_0}}.

For each value of \ifelse{html}{\out{<i>&pi;</i>}}{\eqn{\pi}} in
the supplied vector \ifelse{html}{\out{<b><i>&pi;</i></b>}}{\eqn{\bold{\pi}}},
\code{opchar_gs()} evaluates the power, ESS, and other key characteristics,
of each of the supplied designs.

Calculations are performed conditional on the trial stopping in one of the
stages specified using the input (vector) \code{k}.
}
\examples{
# Find the optimal two-stage design for the default parameters
des         <- des_gs()
# Determine operating characteristics for a range of response probabilities
opchar      <- opchar_gs(des, pi = seq(0, 1, 0.01))
# Find the optimal two-stage design for a 10\% type-I error rate
des_10      <- des_gs(alpha = 0.1)
# Determine operating characteristics for both designs at the same time
opchar_both <- opchar_gs(des, des_10, pi = seq(0, 1, 0.01))
}
\references{
Simon R (1989) Optimal two-stage designs for phase II clinical
trials. \emph{Controlled Clinical Trials} \strong{10:}1-10.
}
\seealso{
\code{\link{des_gs}}, \code{\link{est_gs}}, \code{\link{pval_gs}},
\code{\link{ci_gs}}, and their associated \code{plot} family of functions.
}
