% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sa_ci_fixed.R
\name{plot.sa_ci_fixed}
\alias{plot.sa_ci_fixed}
\title{Plot the confidence intervals, and the confidence interval determination
procedures performance, in a single-stage single-arm trial design for a
single binary endpoint}
\usage{
\method{plot}{sa_ci_fixed}(x, ..., output = F)
}
\arguments{
\item{x}{An object of class \code{"sa_ci_fixed"}, as returned by
\code{ci_fixed()}.}

\item{...}{Included for compatibility with the generic. Not currently used.}

\item{output}{A logical variable indicating whether the outputs described
below should be returned.}
}
\value{
If \code{output = TRUE}, a list containing the following elements is
returned
\itemize{
\item A list in the slot \code{$plot_des} containing the available plots.
\item Each of the input variables as specified, subject to internal
modification.
}
}
\description{
Plots the confidence intervals, and the performance of the confidence
interval determination procedures, in a single-stage single-arm trial design
for a single binary endpoint determined using \code{ci_fixed()}. A range of
plots are available, of which the confidence intervals and the coverage
probability curve will be printed by default.
}
\examples{
# Find the optimal single-stage design for the default parameters
des <- des_fixed()
# Determine the performance of the confidence interval determination
# procedures for a range of possible response probabilities
ci  <- ci_fixed(des, pi = seq(from = 0, to = 1, by = 0.01))
# Plot the confidence intervals and the determination procedures performance
plot(ci)
}
\seealso{
\code{\link{des_fixed}}, \code{\link{opchar_fixed}},
\code{\link{est_fixed}}, \code{\link{pval_fixed}}, \code{\link{ci_fixed}},
and their associated \code{plot} family of functions.
}
