% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sa_des_adaptive.R
\name{plot.sa_des_adaptive}
\alias{plot.sa_des_adaptive}
\title{Plot the second stage sample sizes of two-stage adaptive single-arm trial
designs for a single binary endpoint}
\usage{
\method{plot}{sa_des_adaptive}(x, ..., output = F)
}
\arguments{
\item{x}{An object of class \code{"sa_des_adaptive"}, as returned by \code{des_adaptive()}.}

\item{...}{Additional objects of class \code{"sa_des_adaptive"}. These will be grouped
in to a list named \code{"add_des"}.}

\item{output}{A logical variable indicating whether the outputs described below
should be returned.}
}
\value{
If \code{output = TRUE}, a list containing the following elements is returned
\itemize{
\item A list in the slot \code{$plot_des} containing the available plots.
\item Each of the input variables as specified, subject to internal modification.
}
}
\description{
Plots the second stage sample sizes of two-stage adaptive single-arm trial
designs determined using \code{des_adaptive()}. The values of
\ifelse{html}{\out{<i>n</i><sub>2</sub>(<i>s</i><sub>n<sub>1</sub>
</sub>))}}{\eqn{n_2(\tilde{s}_1)}} are plotted.
}
\details{
Support is available to simultaneously plot the second stage sample sizes of multiple
two-stage adaptive single-arm clinical trial designs for a single binary primary
endpoint.
}
\examples{
# Find the optimal adaptive two-stage design for the default parameters
des    <- des_adaptive()
# Plot the stopping boundaries
plot(des)
# Find the optimal adaptive two-stage design for a 10\% type-I error rate
des_10 <- des_adaptive(alpha = 0.1)
# Plot the second stage sample sizes for both designs
plot(des, des_10)
}
\seealso{
\code{\link{des_adaptive}}, \code{\link{opchar_adaptive}}, and their
associated \code{plot} family of functions.
}
