% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sa_des_admissable.R
\name{plot.sa_des_admissable}
\alias{plot.sa_des_admissable}
\title{Plot the stopping boundaries of admissable group sequential single-arm trial
designs for a single binary endpoint}
\usage{
\method{plot}{sa_des_admissable}(des, output = F)
}
\arguments{
\item{des}{An object of class \code{"sa_des_admissable"}, as returned by
\code{des_admissable()}.}

\item{output}{A logical variable indicating whether the outputs described
below should be returned.}
}
\value{
If \code{output = TRUE}, a list containing the following elements is
returned
\itemize{
\item A list in the slot \code{$plot_des} containing the available plots.
\item Each of the input variables as specified, subject to internal
modification.
}
}
\description{
Plots the stopping boundaries of admissable group sequential single-arm trial
designs determined using \code{des_admissable()}. The possible
\ifelse{html}{\out{(<i>s</i>,<i>m</i>)}}{\eqn{(s,m)}} states during the trial
are plotted in a colour coded manner to indicate their associated decision
rules.
}
\details{
In addition, admissable design plots are produced to depict in which region
of the weight-space each of the designs is optimal, and to illuminate the
expected sample size characteristics of the admissable designs.
}
\examples{
# Find the admissable two-stage designs for the default parameters
des <- des_admissable()
# Plot the stopping boundaries and admissable design plot
plot(des)
}
\seealso{
\code{\link{des_admissable}}, \code{\link{opchar_admissable}},
and their associated \code{plot} family of functions.
}
