% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sa_des_fixed.R
\name{plot.sa_des_fixed}
\alias{plot.sa_des_fixed}
\title{Plot the stopping boundaries of single-stage single-arm trial designs for a
single binary endpoint}
\usage{
\method{plot}{sa_des_fixed}(x, ..., output = F)
}
\arguments{
\item{x}{An object of class \code{"sa_des_fixed"}, as returned by
\code{des_fixed()}.}

\item{...}{Additional objects of class \code{"sa_des_fixed"}. These will be
grouped in to a list named \code{"add_des"}.}

\item{output}{A logical variable indicating whether the outputs described
below should be returned.}
}
\value{
If \code{output = TRUE}, a list containing the following elements is
returned
\itemize{
\item A list in the slot \code{$plot_des} containing the available plots.
\item A tibble in the slot \code{$states} containing details of the possible
states for each of the designs, and their associated decision rules.
\item Each of the input variables as specified, subject to internal
modification.
}
}
\description{
Plots the stopping boundaries of single-stage single-arm trial designs
determined using \code{des_fixed()}. The possible
\ifelse{html}{\out{(<i>s</i>,<i>m</i>)}}{\eqn{(s,m)}} states during the trial
are plotted in a colour coded manner, to indicate their associated decision
rules.
}
\details{
Support is available to simultaneously plot the stopping boundaries of
multiple applicable designs, using faceting.
}
\examples{
# Find the optimal single-stage design for the default parameters
des    <- des_fixed()
# Plot the stopping boundaries
plot(des)
# Find the optimal single-stage design for a 10\% type-I error rate
des_10 <- des_fixed(alpha = 0.1)
# Plot the stopping boundaries for both designs
plot(des, des_10)
}
\seealso{
\code{\link{des_fixed}}, \code{\link{opchar_fixed}},
\code{\link{est_fixed}}, \code{\link{pval_fixed}}, \code{\link{ci_fixed}},
and their associated \code{plot} family of functions.
}
