% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sa_des_gs.R
\name{plot.sa_des_gs}
\alias{plot.sa_des_gs}
\title{Plot the stopping boundaries of group sequential single-arm trial designs for a
single binary endpoint}
\usage{
\method{plot}{sa_des_gs}(des, ..., output = F)
}
\arguments{
\item{des}{An object of class \code{"sa_des_gs"}, as returned by \code{des_gs()}.}

\item{...}{Additional objects of class \code{"sa_des_gs"}. These will be grouped
in to a list named \code{"add_des"}.}

\item{output}{A logical variable indicating whether the outputs described below
should be returned.}
}
\value{
If \code{output = TRUE}, a list containing the following elements is returned
\itemize{
\item A list in the slot \code{$plot_des} containing the available plots.
\item Each of the input variables as specified, subject to internal modification.
}
}
\description{
Plots the stopping boundaries of group sequential single-arm trial designs
determined using \code{des_gs()}. The possible
\ifelse{html}{\out{(<i>s</i>,<i>m</i>)}}{\eqn{(s,m)}} states during the trial
are plotted in a colour coded manner to indicate their associated decision rules.
}
\details{
Support is available to simultaneously plot the stopping boundaries of multiple
group sequential single-arm clinical trial designs for a single binary primary
endpoint, using faceting.
}
\examples{
# Find the optimal two-stage design for the default parameters
des         <- des_gs()
# Plot the stopping boundaries
plot(des)
# Find the optimal two-stage design for a 10\% type-I error rate
des_10      <- des_gs(alpha = 0.1)
# Plot the stopping boundaries for both designs
plot(des, des_10)
}
\seealso{
\code{\link{des_gs}}, \code{\link{opchar_gs}}, \code{\link{est_gs}},
\code{\link{pval_gs}}, \code{\link{ci_gs}}, and their associated \code{plot}
family of functions.
}
