% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sa_est_curtailed.R
\name{plot.sa_est_curtailed}
\alias{plot.sa_est_curtailed}
\title{Plot the point estimates, and the point estimation procedures performance, in
a curtailed group sequential single-arm trial design for a single binary
endpoint}
\usage{
\method{plot}{sa_est_curtailed}(x, ..., output = F)
}
\arguments{
\item{x}{An object of class \code{"sa_est_curtailed"}, as returned by
\code{est_curtailed()}.}

\item{...}{Included for compatibility with the generic. Not currently used.}

\item{output}{A logical variable indicating whether the outputs described
below should be returned.}
}
\value{
If \code{output = TRUE}, a list containing the following elements is
returned
\itemize{
\item A list in the slot \code{$plot_des} containing the available plots.
\item Each of the input variables as specified, subject to internal
modification.
}
}
\description{
Plots the point estimates, and the performance of the point estimation
procedure, in a curtailed group sequential single-arm trial design for a
single binary endpoint determined using \code{est_curtailed()}. A range of
plots are available, of which the bias curve will be printed by default.
}
\examples{
# Find the optimal curtailed group sequential design for the default
parameters
des <- des_curtailed()
# Determine the performance of the point estimation procedure for a range of
# possible response probabilities
est <- est_curtailed(des, pi = seq(0, 1, 0.01))
# Plot the point estimates and the estimation procedure performance
plot(est)
}
\seealso{
\code{\link{des_curtailed}}, \code{\link{opchar_curtailed}},
\code{\link{est_curtailed}}, \code{\link{pval_curtailed}},
\code{\link{ci_curtailed}}, and their associated \code{plot} family of
functions.
}
