% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sa_est_gs.R
\name{plot.sa_est_gs}
\alias{plot.sa_est_gs}
\title{Plot the point estimates, and the point estimation procedures performance, in a group sequential
single-arm trial design for a single binary endpoint}
\usage{
\method{plot}{sa_est_gs}(est, output = F)
}
\arguments{
\item{est}{An object of class \code{"sa_est_gs"}, as returned by \code{est_gs()}.}

\item{output}{A logical variable indicating whether the outputs described below
should be returned.}
}
\value{
If \code{output = TRUE}, a list containing the following elements is returned
\itemize{
\item A list in the slot \code{$plot_des} containing the available plots.
\item Each of the input variables as specified, subject to internal modification.
}
}
\description{
Plots the point estimates, and the performance of the point estimation procedure,
in a group sequential single-arm trial design for a single binary endpoint
determined using \code{est_gs()}. A range of plots are available, of which
the point estimates and the bias curve will be printed by
default.
}
\examples{
# Find the optimal group sequential design for the default parameters
des <- des_gs()
# Determine the performance of the point estimation procedure for a range of
# possible response probabilities
est <- est_gs(des, pi = seq(0, 1, 0.01))
# Plot the point estimates and the estimation procedure performance
plot(est)
}
\seealso{
\code{\link{des_gs}}, \code{\link{opchar_gs}}, \code{\link{est_gs}},
\code{\link{pval_gs}}, \code{\link{ci_gs}}, and their associated \code{plot}
family of functions.
}
