% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sa_opchar_adaptive.R
\name{plot.sa_opchar_adaptive}
\alias{plot.sa_opchar_adaptive}
\title{Plot the operating characteristics of adaptive two-stage single-arm trial designs for a
single binary endpoint}
\usage{
\method{plot}{sa_opchar_adaptive}(x, ..., output = F)
}
\arguments{
\item{x}{An object of class \code{"sa_opchar_adaptive"}, as returned by \code{opchar_adaptive()}.}

\item{...}{Included for compatibility with the generic. Not currently used.}

\item{output}{A logical variable indicating whether the outputs described below
should be returned.}
}
\value{
If \code{output = TRUE}, a list containing the following elements is returned
\itemize{
\item A list in the slot \code{$plot_des} containing the available plots.
\item Each of the input variables as specified, subject to internal modification.
}
}
\description{
Plots the operating characteristics of adaptive two-stage single-arm trial designs
determined using \code{opchar_adaptive()}. A range of plots are available, of which
the expected sample size and power curves will be printed by default.
}
\examples{
# Find the optimal group sequential design for the default parameters
des         <- des_adaptive()
# Determine operating characteristics for a range of response probabilities
opchar      <- opchar_adaptive(des, pi = seq(0, 1, 0.01))
# Plot the expected sample size and power curves
plot(opchar)
}
\seealso{
\code{\link{des_adaptive}}, \code{\link{opchar_adaptive}}, and their
associated \code{plot} family of functions.
}
