% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sa_opchar_curtailed.R
\name{plot.sa_opchar_curtailed}
\alias{plot.sa_opchar_curtailed}
\title{Plot the operating characteristics of curtailed group sequential single-arm trial designs for a
single binary endpoint}
\usage{
\method{plot}{sa_opchar_curtailed}(opchar, output = F)
}
\arguments{
\item{opchar}{An object of class \code{"sa_opchar_curtailed"}, as returned by \code{opchar_curtailed()}.}

\item{output}{A logical variable indicating whether the outputs described below
should be returned.}
}
\value{
If \code{output = TRUE}, a list containing the following elements is returned
\itemize{
\item A list in the slot \code{$plot_des} containing the available plots.
\item Each of the input variables as specified, subject to internal modification.
}
}
\description{
Plots the operating characteristics of curtailed group sequential single-arm trial designs
determined using \code{opchar_curtailed()}. A range of plots are available, of which
the expected sample size and power curves will be printed by default.
}
\examples{
# Find the optimal group sequential design for the default parameters with NSC
des         <- des_curtailed()
# Determine operating characteristics for a range of response probabilities
opchar      <- opchar_curtailed(des, pi = seq(0, 1, 0.01))
# Plot the expected sample size and power curve
plot(opchar)
}
\seealso{
\code{\link{des_curtailed}}, \code{\link{opchar_curtailed}},
and their associated \code{plot} family of functions.
}
