% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sa_pval_fixed.R
\name{plot.sa_pval_fixed}
\alias{plot.sa_pval_fixed}
\title{Plot the p-values, and the p-value calculation procedures performance, in a
single-stage single-arm trial design for a single binary endpoint}
\usage{
\method{plot}{sa_pval_fixed}(x, ..., output = F)
}
\arguments{
\item{x}{An object of class \code{"sa_pval_fixed"}, as returned by
\code{pval_fixed()}.}

\item{...}{Included for compatibility with the generic. Not currently used.}

\item{output}{A logical variable indicating whether the outputs described
below should be returned.}
}
\value{
If \code{output = TRUE}, a list containing the following elements is
returned
\itemize{
\item A list in the slot \code{$plot_des} containing the available plots.
\item Each of the input variables as specified, subject to internal
modification.
}
}
\description{
Plots the p-values, and the performance of the p-value calculation
procedures, in a single-stage single-arm trial design for a single binary
endpoint determined using \code{pval_fixed()}. A range of plots are
available, of which the p-values and the expected p-value curve will be
printed by default.
}
\examples{
# Find the optimal single-stage design for the default parameters
des  <- des_fixed()
# Determine the performance of the p-value calculation procedures for a range
# of possible response probabilities
pval <- pval_fixed(des, pi = seq(from = 0, to = 1, by = 0.01))
# Plot the p-values and the calculation procedures performance
plot(pval)
}
\seealso{
\code{\link{des_fixed}}, \code{\link{opchar_fixed}},
\code{\link{est_fixed}}, \code{\link{pval_fixed}}, \code{\link{ci_fixed}},
and their associated \code{plot} family of functions.
}
