% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sa_pval_gs.R
\name{plot.sa_pval_gs}
\alias{plot.sa_pval_gs}
\title{Plot the p-values, and the p-value calculation procedures performance, in a group sequential
single-arm trial design for a single binary endpoint}
\usage{
\method{plot}{sa_pval_gs}(pval, output = F)
}
\arguments{
\item{pval}{An object of class \code{"sa_pval_gs"}, as returned by \code{pval_gs()}.}

\item{output}{A logical variable indicating whether the outputs described below
should be returned.}
}
\value{
If \code{output = TRUE}, a list containing the following elements is returned
\itemize{
\item A list in the slot \code{$plot_des} containing the available plots.
\item Each of the input variables as specified, subject to internal modification.
}
}
\description{
Plots the p-values, and the performance of the p-value calculation procedures,
in a group sequential single-arm trial design for a single binary endpoint
determined using \code{pval_gs()}. A range of plots are available, of which
the p-values and the expected p-value curve will be printed by default.
}
\examples{
# Find the optimal group sequential design for the default parameters
des  <- des_gs()
# Determine the performance of the p-value calculation procedures for a range of
# possible response probabilities
pval <- pval_gs(des, pi = seq(from = 0, to = 1, by = 0.01))
# Plot the p-values and the calculation procedures performance
plot(pval)
}
\seealso{
\code{\link{des_gs}}, \code{\link{opchar_gs}}, \code{\link{est_gs}}, \code{\link{pval_gs}},
and \code{\link{ci_gs}}, and their associated \code{plot} family of functions.
}
