% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_fixed.R
\name{pval_fixed}
\alias{pval_fixed}
\title{Determine p-values in a single-stage single-arm trial design for a single
binary endpoint}
\usage{
pval_fixed(des, pi, method = c("exact", "normal"), summary = F)
}
\arguments{
\item{des}{An object of class \code{"sa_des_fixed"}, as returned by
\code{des_fixed()}.}

\item{pi}{A vector of response probabilities to evaluate the expected
performance of the p-value calculation procedure at. This will internally
default to be the \ifelse{html}{\out{<i>&pi;</i><sub>0</sub>}}{\deqn{\pi_0}}
and \ifelse{html}{\out{<i>&pi;</i><sub>1</sub>}}{\deqn{\pi_1}} from
\code{des} if it is left unspecified.}

\item{method}{A vector of methods to use to construct p-values. Currently,
support is available to use exact binomial tail probabilities
(\code{"exact"}) and to use a normal approximation (\code{"normal"})
approach.}

\item{summary}{A logical variable indicating whether a summary of the
function's progress should be printed to the console.}
}
\value{
A list of class \code{"sa_pval_fixed"} containing the following
elements
\itemize{
\item A tibble in the slot \code{$pval} summarising the possible p-values at
the end of the trial for the supplied design, according to the chosen
methods.
\item A tibble in the slot \code{$perf} summarising the performance of the
chosen p-value calculation procedures for each specified value of
\ifelse{html}{\out{<i>&pi;</i>}}{\deqn{\pi}}.
\item Each of the input variables as specified, subject to internal
modification.
}
}
\description{
Determines all possible p-values at the end of a single-stage single-arm
trial for a single binary endpoint, as determined using \code{des_fixed()}.
Support is available to compute p-values using exact binomial tail
probabilities, or using a normal approximation.
}
\details{
In addition, the performance of the chosen p-value calculation procedures
(including their expected value and variance) for each value of
\ifelse{html}{\out{<i>pi</i>}}{\eqn{\pi}} in the supplied vector
\ifelse{html}{\out{<b><i>pi</i></b>}}{\eqn{\bold{\pi}}}, will also be
evaluated.
}
\examples{
# Find the optimal single-stage design for the default parameters
des  <- des_fixed()
# Determine the performance of both supported p-value calculation procedures
# for a range of possible response probabilities
pval <- pval_fixed(des, pi = seq(0, 1, 0.01))
}
\seealso{
\code{\link{des_fixed}}, \code{\link{opchar_fixed}},
\code{\link{est_fixed}}, \code{\link{ci_fixed}}, and their associated
\code{plot} family of functions.
}
